+++ R/PredictionData.R
@@ -22,7 +22,7 @@ NULL
 
 new_prediction_data = function(li, task_type) {
   li = discard(li, is.null)
-  class(li) = c(fget(mlr_reflections$task_types, task_type, "prediction_data", "type"), "PredictionData")
+  class(li) = c(fget_key(mlr_reflections$task_types, task_type, "prediction_data", "type"), "PredictionData")
   li
 }