+++ R/benchmark.R
@@ -6,6 +6,10 @@
 #' For large-scale benchmarking we recommend to use the \CRANpkg{mlr3batchmark} package.
 #' This package runs benchmark experiments on high-performance computing clusters and handles failed experiments.
 #'
+#' @section Stochasticity:
+#' Note that uninstantiated [`Resampling`]s are instantiated on the task, making
+#' the function stochastic even in case of deterministic learners.
+#'
 #' @param design ([data.frame()])\cr
 #'   Data frame (or [data.table::data.table()]) with three columns: "task", "learner", and "resampling".
 #'   Each row defines a resampling by providing a [Task], [Learner] and an instantiated [Resampling] strategy.