+++ NEWS.md
@@ -1,5 +1,9 @@
 # mlr3 (development version)
 
+# mlr3 1.1.0
+
+* feat: Add new measure `MeasureRegrRQR` for quantile regression.
+* feat: Add `$predict_newdata_fast()` method to `Learner` to speed up prediction.
 * fix: `configure_learner` is passed on `run_experiment()` for autotest learners.
 
 # mlr3 1.0.1